+++ packages/database/src/gather/entities/gather-state.entity.ts
@@ -1,8 +1,18 @@
 import { BaseEntity } from '../../core/base.entity.js';
-import { ExpectedEventState } from '../types/expected-event-state.interface.js';
+import { GatherEventState } from '../types/gather-event-state.interface.js';
+import type { EventMessagingType, EventRegistry } from '@volontariapp/messaging';
 
-export class GatherStateEntity extends BaseEntity {
+export class GatherStateMetadata<TKey extends EventMessagingType = EventMessagingType> {
+  emitterId?: string;
+  traceId?: string;
+  payload?: EventRegistry[TKey];
+}
+
+export class GatherStateEntity<
+  TKey extends EventMessagingType = EventMessagingType,
+> extends BaseEntity {
   correlationId!: string;
-  triggerEvent!: string;
-  eventsState: Record<string, ExpectedEventState> = {};
+  triggerEvent!: TKey;
+  gatherEventsState: Record<string, GatherEventState> = {};
+  metadata?: GatherStateMetadata<TKey>;
 }