+++ packages/database/src/gather/types/index.ts
@@ -1,2 +1,2 @@
 export * from './event-status.enum.js';
-export * from './expected-event-state.interface.js';
+export * from './gather-event-state.interface.js';