+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.2.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/database@3.4.1
+  - @volontariapp/outbox@0.9.38
+
 ## 3.2.4
 
 ### Patch Changes